refactor: extract vault_db primitive, tighten SSH narrative
Split the vault-access context manager into its own domain primitive
(lib/domain/vault.py) so sibling domain modules can use it without
reaching up to the facade — a cleaner layering and a clearer narrative
chapter in the manifesto sense.

Also picks up the terok-sandbox tip that splits SSHManager's
constructor into caller-owned (__init__) and context-managed (open)
variants, and trims a four-line docstring on ssh_use_personal to one
line that states the contract.
