+++ src/terok/lib/domain/project_state.py
@@ -21,7 +21,7 @@ from ..core.task_display import container_name as _container_name
 
 def _scope_has_vault_key(scope: str) -> bool:
     """Return ``True`` iff the vault has at least one SSH key assigned to *scope*."""
-    from .facade import vault_db
+    from .vault import vault_db
 
     with vault_db() as db:
         return bool(db.list_ssh_keys_for_scope(scope))
